# docs/openapi/mediflow-v1.yaml
openapi: 3.1.0
info:
  title: MediFlow Local API v1
  version: 1.16.0
  summary: Stable patient contract plus executable home-base network slice with explicit pairing review, paired-client auth boundary, ambulatory scope options, scoped remote patient, diary, therapy, checkup, observation, and catalog read boundaries, replica planning state, explicit identity summary, and declarative AI runtime summary for the local `/api/v1` surface.
  description: |
    OpenAPI contract for the client-facing local API defined by ADR 0010.
    Current thin slice covers stable patient reads plus compare-on-write patient
    mutations on `PUT /api/v1/patients/{id}` and `DELETE /api/v1/patients/{id}`.
    It also introduces the first documented `/api/v1/network/*` contract slice
    for home-base discovery, session gating, capability discovery, explicit
    pairing bootstrap/review on trusted local networks, the first read-only
    remote patient data plane, the first paired patient profile/status write
    boundary, the first paired clinical diary read/create/update boundary, the
    first paired therapy read/create/update/soft-delete boundary, the first
    paired checkup read/create/update/soft-delete boundary, the first paired
    observation read/create/update/soft-delete boundary, and the first paired
    read-only catalog boundary for global drug, exemption, and terminology
    reference data behind dedicated paired-client auth capabilities.
...
servers:
  - url: https://127.0.0.1:3443
...
paths:
...
  /api/v1/network/prosthetic-prescriptions/{id}:
    put:
      tags:
        - Network
      operationId: updateNetworkProstheticPrescription
      summary: Update prosthetic prescription through the paired boundary
      description: |
        Applies a compare-on-write update to a prosthetic prescription in
        scope. `version` is required and stale versions return
        `VERSION_CONFLICT`. This route requires
        `network.replica.write-prosthetic-prescriptions`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Prosthetic prescription identifier.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProstheticPrescriptionMutationRequest'
      responses:
        '200':
          description: Prosthetic prescription updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PrescriptionVersionConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/network/patients:
    get:
      tags:
        - Network
      operationId: listNetworkPatients
      summary: List patients through the paired home-base data plane
      description: |
        Returns read-only patient summaries scoped by the effective ambulatory
        resolved on the home-base node. This route requires both a valid paired
        client credential and an operator session cookie. `includeDeleted=true`
        also requires `network.replica.write-patient-lifecycle` and returns
        tombstoned patients with their deletion metadata.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: includeDeleted
          in: query
          required: false
          description: Include soft-deleted patients when set to `true`.
          schema:
            type: boolean
      responses:
        '200':
          description: Read-only scoped patient summaries from the paired node.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientSummary'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Network
      operationId: createNetworkPatient
      summary: Create patient through the paired lifecycle boundary
      description: |
        Creates a patient and its ambulatory membership atomically from a paired
        client with `network.replica.write-patient-lifecycle`. The insert always
        targets the effective ambulatory scope resolved for the paired session,
        the same resolution used by the other network patient routes; a body
        `ambulatoryId` that differs from that scope is rejected.
        Sensitive optional fields must already be sealed with the `ENC:` format.
        AI/document fields are rejected.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkPatientCreateRequest'
            examples:
              createNetworkPatient:
                value:
                  firstName: Ada
                  lastName: Lovelace
                  taxCode: LVLDAA85A01F205X
                  birthDate: '1985-01-01T00:00:00.000Z'
                  address: ENC:iv:ciphertext
                  notes: ENC:iv:notes
      responses:
        '201':
          description: Patient created through the paired lifecycle boundary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedVersionedResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/network/patients/{id}:
    get:
      tags:
        - Network
      operationId: getNetworkPatient
      summary: Get patient detail through the paired home-base data plane
      description: |
        Returns read-only patient detail from the paired node, limited to the
        effective ambulatory scope resolved by the home-base. This route
        requires both a valid paired client credential and an operator session
        cookie.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Patient identifier.
          schema:
            type: string
      responses:
        '200':
          description: Read-only scoped patient detail from the paired node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDetail'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Network
      operationId: updateNetworkPatient
      summary: Update patient profile/status through the paired home-base data plane
      description: |
        Updates a patient profile/status subset from a paired client using the
        same optimistic concurrency shape as the local patient mutation
        contract. The route requires a paired client with
        `network.replica.write-patient-profile`, an active operator session,
        and the effective ambulatory scope resolved by the home-base node.
        AI/document-derived fields, child clinical CRUD, remote delete, and
        sync queues remain outside this first write boundary.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Patient identifier.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientMutationRequest'
            examples:
              updateNetworkPhone:
                value:
                  version: 7
                  phone: "+39 333 1234567"
              archiveNetworkPatient:
                value:
                  version: 7
                  isArchived: true
      responses:
        '200':
          description: Patient updated through the paired boundary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PatientVersionConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Network
      operationId: deleteNetworkPatient
      summary: Soft-delete patient through the paired lifecycle boundary
      description: |
        Soft-deletes an active patient using optimistic concurrency. Requires
        `network.replica.write-patient-lifecycle`, paired client credentials,
        an active operator session, and membership in the effective ambulatory
        scope. `deletionReason`, when provided, must be sealed with `ENC:`.
      security:
        - pairedClientId: []
          pairedClientToken: []
          sessionCookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Patient identifier.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkPatientDeleteRequest'
            examples:
              deleteNetworkPatient:
                value:
                  version: 7
                  deletionReason: ENC:iv:reason
      responses:
        '200':
          description: Patient soft-deleted through the paired lifecycle boundary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedVersionedResource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/NetworkDataPlaneUnauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/PatientVersionConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
      description: "Local API token expected in `Authorization: Bearer <token>`."
    pairedClientId:
      type: apiKey
      in: header
      name: x-mediflow-paired-client-id
      description: Explicit paired-device identifier minted during local pairing confirmation.
    pairedClientToken:
      type: apiKey
      in: header
      name: x-mediflow-paired-client-token
      description: One-time paired-device token minted during local pairing confirmation.
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: mediflow_session
      description: Operator session cookie required for the remote network data plane.
  responses:
    BadRequest:
      description: Request payload is invalid or missing required mutation metadata.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingVersion:
              value:
                error: Version is required
            noFields:
              value:
                error: No valid fields to update
...
    NetworkDataPlaneUnauthorized:
      description: Missing or invalid paired client credential, or missing operator session cookie.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: Unauthorized
    Forbidden:
      description: Request is authenticated but blocked by missing scope or missing granted capability.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                error: Forbidden
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                error: Not found
...
    PatientVersionConflict:
      description: |
        Compare-on-write conflict. The payload exposes only PHI-safe metadata so
        clients can refresh, inspect the version delta, and decide whether to retry.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PatientVersionConflictResponse'
          examples:
            outdatedVersion:
              value:
                error: Conflict
                code: VERSION_CONFLICT
                entity: patient
                recordId: patient_123
                expectedVersion: 7
                currentVersion: 8
                currentUpdatedAt: '2026-03-16T10:15:00.000Z'
                currentState: present
                currentSnapshot:
                  id: patient_123
                  version: 8
                  updatedAt: '2026-03-16T10:15:00.000Z'
                  isArchived: false
            deletedElsewhere:
              value:
                error: Conflict
                code: VERSION_CONFLICT
                entity: patient
                recordId: patient_123
                expectedVersion: 7
                currentVersion: null
                currentUpdatedAt: null
                currentState: missing
                currentSnapshot: null
...
    PrescriptionVersionConflict:
      description: |
        Prescription compare-on-write conflict. The payload exposes only
        PHI-safe metadata so clients can refresh and decide whether to retry.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PrescriptionVersionConflictResponse'
          examples:
            outdatedVersion:
              value:
                error: Conflict
                code: VERSION_CONFLICT
                entity: service_prescription
                recordId: prescription_123
                expectedVersion: 2
                currentVersion: 3
                currentUpdatedAt: '2026-05-02T10:15:00.000Z'
                currentState: present
                currentSnapshot:
                  id: prescription_123
                  patientId: patient_123
                  version: 3
                  updatedAt: '2026-05-02T10:15:00.000Z'
            missingPrescription:
              value:
                error: Conflict
                code: VERSION_CONFLICT
                entity: service_prescription
                recordId: prescription_123
                expectedVersion: 2
                currentVersion: null
                currentUpdatedAt: null
                currentState: missing
                currentSnapshot: null
    InternalServerError:
      description: Unexpected server-side failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            patientListFailure:
              value:
                error: Failed to fetch patients
            patientDetailFailure:
              value:
                error: Failed to fetch patient
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
...
    PatientSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - firstName
        - lastName
        - taxCode
        - version
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type:
            - string
            - 'null'
          format: date-time
        taxCode:
          type: string
        isAdi:
          type:
            - boolean
            - 'null'
        isArchived:
          type:
            - boolean
            - 'null'
        deletedAt:
          type:
            - string
            - 'null'
          format: date-time
        deletionReason:
          type:
            - string
            - 'null'
        version:
          type: integer
          minimum: 1
        updatedAt:
          type:
            - string
            - 'null'
          format: date-time
    PatientDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - firstName
        - lastName
        - taxCode
        - version
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type:
            - string
            - 'null'
          format: date-time
        taxCode:
          type: string
        address:
          type:
            - string
            - 'null'
        phone:
          type:
            - string
            - 'null'
        caregiver:
          type:
            - string
            - 'null'
        exemptions:
          type:
            - string
            - 'null'
        diagnoses:
          type:
            - string
            - 'null'
        monitoringProfile:
          type:
            - string
            - 'null'
        statusReason:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
        aiSummary:
          type:
            - string
            - 'null'
        aiSummaryGeneratedAt:
          type:
            - string
            - 'null'
          format: date-time
        aiSummaryContextHash:
          type:
            - string
            - 'null'
        documentInsights:
          type:
            - string
            - 'null'
        isAdi:
          type:
            - boolean
            - 'null'
        isArchived:
          type:
            - boolean
            - 'null'
        deletedAt:
          type:
            - string
            - 'null'
          format: date-time
        deletionReason:
          type:
            - string
            - 'null'
        version:
          type: integer
          minimum: 1
        ambulatoryId:
          type:
            - string
            - 'null'
        createdAt:
          type:
            - string
            - 'null'
          format: date-time
        updatedAt:
          type:
            - string
            - 'null'
          format: date-time
...
    PatientMutationRequest:
      type: object
      additionalProperties: false
      required:
        - version
      description: |
        Partial patient update payload. At least one field other than `version`
        must be provided.
      properties:
        version:
          type: integer
          minimum: 1
        firstName:
          type: string
        lastName:
          type: string
        taxCode:
          type: string
        birthDate:
          type:
            - string
            - 'null'
          format: date-time
        address:
          type:
            - string
            - 'null'
        phone:
          type:
            - string
            - 'null'
        caregiver:
          type:
            - string
            - 'null'
        exemptions:
          type:
            - string
            - 'null'
        diagnoses:
          type:
            - string
            - 'null'
        monitoringProfile:
          type:
            - string
            - 'null'
        statusReason:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
        aiSummary:
          type:
            - string
            - 'null'
        aiSummaryGeneratedAt:
          type:
            - string
            - 'null'
          format: date-time
        aiSummaryContextHash:
          type:
            - string
            - 'null'
        documentInsights:
          type:
            - string
            - 'null'
        isAdi:
          type: boolean
        isArchived:
          type: boolean
        ambulatoryId:
          type:
            - string
            - 'null'
    NetworkPatientCreateRequest:
      type: object
      additionalProperties: false
      required:
        - firstName
        - lastName
        - taxCode
      description: |
        Patient create payload for paired lifecycle writes. Sensitive optional
        fields must be sealed strings beginning with `ENC:`. Server-controlled
        fields and AI/document fields are rejected.
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        taxCode:
          type: string
        birthDate:
          type:
            - string
            - 'null'
          format: date-time
        address:
          type:
            - string
            - 'null'
        phone:
          type:
            - string
            - 'null'
        caregiver:
          type:
            - string
            - 'null'
        exemptions:
          type:
            - string
            - 'null'
        diagnoses:
          type:
            - string
            - 'null'
        monitoringProfile:
          type:
            - string
            - 'null'
        statusReason:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
        isAdi:
          type: boolean
        ambulatoryId:
          type:
            - string
            - 'null'
...
    NetworkPatientDeleteRequest:
      allOf:
        - $ref: '#/components/schemas/PatientDeleteRequest'
        - type: object
          additionalProperties: false
          properties:
            deletionReason:
              type: string
              description: Sealed optional deletion reason beginning with `ENC:`.
    CreatedVersionedResource:
      type: object
      additionalProperties: false
      required:
        - id
        - version
      properties:
        id:
          type: string
        version:
          type: integer
          minimum: 1
    MutationSuccessResponse:
      type: object
      additionalProperties: false
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
...
    ProstheticPrescriptionMutationRequest:
      type: object
      additionalProperties: false
      required:
        - version
      properties:
        version:
          type: integer
          minimum: 1
        prescribedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, prescribed, submitted, authorized, delivered, tested, cancelled]
        category:
          type: string
          enum: [standard, oxygen, repair, replacement, trial, other]
        isoCode:
          type:
            - string
            - 'null'
        description:
          type: string
        measures:
          type:
            - string
            - 'null'
        clinicalReason:
          type:
            - string
            - 'null'
        regionalPrescriptionId:
          type:
            - string
            - 'null'
        supplier:
          type:
            - string
            - 'null'
        collaudoAt:
          type:
            - string
            - 'null'
          format: date-time
        collaudoOutcome:
          type:
            - string
            - 'null'
        source:
          type: string
          enum: [manual, document_review]
        documentRefs:
          type:
            - string
            - 'null'
        notes:
          type:
            - string
            - 'null'
...
    PrescriptionVersionConflictResponse:
      type: object
      additionalProperties: false
      required:
        - error
        - code
        - entity
        - recordId
        - expectedVersion
        - currentVersion
        - currentUpdatedAt
        - currentState
        - currentSnapshot
      properties:
        error:
          type: string
          const: Conflict
        code:
          type: string
          const: VERSION_CONFLICT
        entity:
          type: string
          enum:
            - service_prescription
            - service_prescription_item
            - prosthetic_prescription
        recordId:
          type: string
        expectedVersion:
          type: integer
          minimum: 1
        currentVersion:
          type:
            - integer
            - 'null'
          minimum: 1
        currentUpdatedAt:
          type:
            - string
            - 'null'
          format: date-time
        currentState:
          type: string
          enum:
            - present
            - missing
        currentSnapshot:
          oneOf:
            - $ref: '#/components/schemas/PrescriptionConflictSnapshot'
            - type: 'null'
...